Move connection failure tracking code into a module
