+++ index.js
@@ -8,6 +8,7 @@ var error = require('./lib/error');
 var util = require('./lib/util');
 var provider = require('./lib/provider');
 var wildcard = require('wildcard');
+var failureTracking = require('./lib/failureTracking');
 
 var OPTIONAL_MONITOR_EVENTS = [
     'negotiate:request', 'negotiate:renegotiate', 'negotiate:abort',
@@ -72,33 +73,6 @@ module.exports = function(qc, opts) {
     });   
   }
 
-  var doomCountdowns = {};
-  var haveEndedCandidates = {};
-  var haveCompletedGather = {};
-
-  function checkDoomCountdown(peerId) {
-    if (doomCountdowns[peerId]) {
-      return;
-    }
-
-    if (haveEndedCandidates[peerId] && haveCompletedGather[peerId]) {
-      doomCountdowns[peerId] = setTimeout(function() {
-        console.log('AAAAAAAAARRRRRGHHH failed to connect to peer %s', peerId);
-        endDoomCountdown(peerId);
-      }, 10 * 1000);
-    }
-  }
-
-  function endDoomCountdown(peerId) {
-    var doom = doomCountdowns[peerId];
-    if (doom) {
-      clearTimeout(doom);
-    }
-    delete doomCountdowns[peerId];
-    delete haveEndedCandidates[peerId];
-    delete haveCompletedGather[peerId];
-  }
-
   /**
     Emit a generic notification event that allows for tapping into the activity that is happening
     within quickconnect
@@ -116,15 +90,15 @@ module.exports = function(qc, opts) {
     log(peerId, pc, data);
     var gatherEvent = 'pc.' + peerId + '.ice.gathercomplete';
     qc.on(gatherEvent, function() {
-      haveCompletedGather[peerId] = true;
-      checkDoomCountdown(peerId);
+      failureTracking.haveCompletedGather[peerId] = true;
+      failureTracking.checkDoomCountdown(peerId);
     });
     return tc;
   }
 
   qc.on('message:endofcandidates', function(msg, peer, raw) {
-    haveEndedCandidates[peer.id] = true;
-    checkDoomCountdown(peer.id);
+    failureTracking.haveEndedCandidates[peer.id] = true;
+    failureTracking.checkDoomCountdown(peer.id);
   });
 
   /**
